LayoutTests/platform/chromium/plugins/drag-events.html

<html>
<head>
<style>
  #plugin {
    width: 150px;
    height: 150px;
  }
</style>
</head>

<body onload="test();">
<embed id="plugin" type="application/x-webkit-test-webplugin"></embed>
<div contenteditable id='text'><span>Some text to drag</span></div>
<script>

function test() {
    if (!window.testRunner || !window.eventSender) {
        document.write("This test does not work in manual mode.");
    } else {
        testRunner.dumpAsText();

        var div = document.getElementById('text');
        div.focus();
        getSelection().collapse(div, 0);
        getSelection().modify('extend', 'forward', 'word');

        var positionX = plugin.offsetLeft + 10;
        var positionY = plugin.offsetTop + 10;

        var startX = div.offsetLeft + div.firstChild.offsetWidth / 5;
        var startY = div.offsetTop + div.offsetHeight / 2;;

        eventSender.dragMode = true;
        eventSender.mouseMoveTo(startX, startY);
        eventSender.mouseDown();
        eventSender.leapForward(250);
        eventSender.mouseMoveTo(positionX, positionY);
        eventSender.mouseMoveTo(positionX + 2, positionY);
        eventSender.mouseMoveTo(startX, startY);
        eventSender.mouseUp();
    }
}
</script>
</body>
</html>
