LayoutTests/platform/chromium/plugins/drag-events-expected.txt

Plugin received event: MouseEnter
Plugin received event: MouseMove
Plugin received event: DragEnter
Plugin received event: DragOver
Plugin received event: DragLeave

Some text to drag
